# Conflicts:
#	src/base/cubstream.hpp
#	src/communication/network_sr.c
#	src/communication/stream_transfer_receiver.cpp
#	src/communication/stream_transfer_sender.cpp
#	src/connection/server_support.c
#	src/executables/master_heartbeat.c
#	src/executables/master_heartbeat.h
#	src/transaction/boot_sr.c
#	unit_tests/CMakeLists.txt
#	unit_tests/transfer_channel/test_main.cpp

+++ src/communication/communication_channel.cpp
@@ -76,7 +76,7 @@ css_error_code communication_channel::send (const std::string &message)
 
 css_error_code communication_channel::recv (char *buffer, std::size_t &maxlen_in_recvlen_out)
 {
-  int copy_of_maxlen_in_recvlen_out = maxlen_in_recvlen_out;
+  int copy_of_maxlen_in_recvlen_out = (int) maxlen_in_recvlen_out;
   int rc = NO_ERRORS;
 
   rc = css_net_recv (m_socket, buffer, &copy_of_maxlen_in_recvlen_out, m_max_timeout_in_ms);
@@ -90,8 +90,8 @@ css_error_code communication_channel::send (const char *buffer, std::size_t leng
   int total_len = 0, rc = NO_ERRORS;
   struct iovec iov[2];
 
-  css_set_io_vector (&iov[0], &iov[1], buffer, length, &templen);
-  total_len = sizeof (int) + length;
+  css_set_io_vector (&iov[0], &iov[1], buffer, (int) length, &templen);
+  total_len = (int) (sizeof (int) + length);
 
   rc = css_send_io_vector_with_socket (m_socket, iov, total_len, vector_length, m_max_timeout_in_ms);
   return (css_error_code) rc;
